Add Code for Romania
Founded in 2016 by a group of young professionals who thought there is a lot more that they can do for their country, Code for Romania is a not-for-profit organisation which aims to empower IT-savvy volunteers to create digital tools that help solve societal problems.
By partnering with NGOs and Public Institutions, we bring together a community of like-minded volunteers that provides technical solutions for citizens, civil society and public services.
For our volunteers to have an operating platform, we have a core team of 18 members for management, operations, strategy, communication, IT and finance. With an average age of 27 years old, our team is currently located in Bucharest, Cluj-Napoca, Barcelona, Rotterdam, and Amsterdam.
Code for Romania is not politically affiliated and will always stay like this.